cloud-functions/rec-daily-trigger-candidate-gen/requirements.txt
functions-framework==3.*
firebase-admin
google-auth
requests